// src/types.ts
export interface recommendationProps {
	original_name: string;
	adult: boolean;
	backdrop_path: string | null;
	id: number;
	title: string;
	original_language: string;
	original_title: string;
	overview: string;
	poster_path: string;
	media_type: 'tv' | 'movie' | 'person';
	genre_ids: number[];
	popularity: number;
	release_date: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
}
// ...
export interface detailsProps {
	adult: boolean;
	backdrop_path: string;
	belongs_to_collection: object;
	budget: number;
	genres: genresProps[];
	homepage: string;
	id: number;
	imdb_id: string;
	name: string;
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	poster_path: string;
	production_companies: object[];
	production_countries: object[];
	release_date: string;
	revenue: number;
	runtime: number;
	spoken_languages: object[];
	status: string;
	tagline: string;
	title: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
}

export interface creditsProps {
	cast: castProps[];
	crew: object[];
	id: number;
}

export interface castProps {
	adult: boolean;
	gender: 1 | 2;
	id: number;
	known_for_department: string;
	name: string;
	original_name: string;
	popularity: number;
	profile_path: string | null;
	credit_id: string;
	department: string;
	job: string;
}

export interface keywordProps {
	id: number;
	name: string;
}

export interface keywordsProps {
	id: number;
	keywords?: keywordProps[];
	results?: keywordProps[];
}

export interface genresProps {
	id: number;
	name: string;
}

export interface recommendationsProps {
	page: number;
	results: recommendationProps[];
	total_pages: number;
	total_results: number;
}
export interface tvRecommendationsProps {
	page: number;
	results: tvRecommendationProps[];
	total_pages: number;
	total_results: number;
}

export interface SearchResultsProps {
	page: number;
	results: SearchResultProps[];
	total_pages: number;
	total_results: number;
}

export interface reviewsProps {
	id: number;
	results: reviewProps[];
	page: number;
	total_pages: number;
	total_results: number;
}

export interface reviewProps {
	author: string;
	author_details: {
		name: string;
		username: string;
		avatar_path: string;
		rating: number;
	};
	content: string;
	created_at: string;
	id: string;
	updated_at: string;
	url: string;
}

export interface videosProps {
	id: number;
	results: videoProps[];
}

export interface videoProps {
	id: string;
	iso_639_1: string;
	iso_3166_1: string;
	key: string;
	name: string;
	site: string;
	size: number;
	type: string;
	official: boolean;
	published_at: string;
}